+++ migration/20230226113702_songs.sql
@@ -3,7 +3,7 @@
 CREATE TABLE playlist (
   id bigserial primary key,
   name varchar(255) unique not null,
-  duration numeric(3, 2) not null
+  duration float(8) not null
 );
 -- +goose StatementEnd
